% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clean_concentration_annotations}
\alias{clean_concentration_annotations}
\title{clean_concentration_annotations}
\usage{
clean_concentration_annotations(x)
}
\arguments{
\item{x}{A character-vector with a hodgepodge of numerical annotations (e.g., "10", "1_in_100", "0,5"))}
}
\value{
A vector of factors
}
\description{
`clean_concentration_annotations()`
}
\details{
Standarizes concentration-values
}
