% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mabid.R
\name{load_mabid}
\alias{load_mabid}
\title{Load an single-cell MAb-ID-sample}
\usage{
load_mabid(
  signal_hdf5,
  map_hdf5,
  control_hdf5,
  reference,
  bin_size = NULL,
  sample_name = NULL,
  threads = 5,
  blacklist = NULL,
  population_average = F,
  pseudocount = 1,
  min_bin = 100,
  downsample = F,
  downsample_control = NULL
)
}
\arguments{
\item{signal_hdf5}{Path(s) to a hdf5-file (binned or per TTAA).}

\item{map_hdf5}{Path to a matching mappability.hdf5}

\item{control_hdf5}{Path(s) to matching control.hdf5 (typically "nones")}

\item{reference}{A BSgenome-object}

\item{bin_size}{The used bin size of the signal-hdf5}

\item{sample_name}{String to for colnaames(sce). Go for a discriptive one!}

\item{threads}{number of threads to use}

\item{blacklist}{A GRanges-object with regions that should be set to NA due to them being not representative. Also include (peri)centromeric regions!}

\item{population_average}{make a population_average?}

\item{pseudocount}{Default = 1}

\item{min_bin}{skip small chromosomes (usually patches)}

\item{downsample}{Downsample counts of all samples to lowest (can also be a number)}

\item{downsample_control}{Downsample control to a number}
}
\value{
An SCE-object of the RangedSummarizedExperiment-subclass (logcounts = TMM-normalised lfc, normcounts = GLM-residuals)
}
\description{
`load_mabid()` returns a SingleCellExperiment-object.
}
\details{
This is the single-cell loader of AbID-data. It will do normalization on control and
mappability and generate multiple handy tracks and metadata.
}
\examples{
\dontrun{
output_sce <- load_scMAbid(signal_hdf5 = "index_30.ABBC_001.SBC_001.100000.hdf5",
                       control_hdf5 = "index_30.ABBC_001.SBC_007.100000.hdf5",
                       map_hdf5 = 'Homo_sapiens.GRCh37.dna.primary_assembly.TTAA.100000.hdf5',
                       bin_size = 100000, reference = BSgenome.Hsapiens.UCSC.hg19)
}
}
