% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mabid_annotate}
\alias{mabid_annotate}
\title{mabid_annotate}
\usage{
mabid_annotate(sce, anno_vec, name, where = NULL)
}
\arguments{
\item{sce}{The mabid-sce}

\item{anno_vec}{A named vector}

\item{name}{The name of the new annotation-colun}

\item{where}{Can be set to match to a column-name in coldata, instead of colnames(sce)}
}
\value{
a new sce
}
\description{
`mabid_annotate()` lets you add a named vector to colData
}
\details{
Easiest why to add some extra annotations
}
