% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabid_tools.R
\name{mabid_bind}
\alias{mabid_bind}
\title{mabid_bind}
\usage{
mabid_bind(sce_1, sce_2)
}
\arguments{
\item{sce_1}{A SCE object}

\item{sce_2}{A SCE object}
}
\value{
An SCE-object of the RangedSummarizedExperiment-subclass (logcounts = TMM-normalised lfc)
}
\description{
`mabid_bind()` binds TTAA and GATC SCE's
}
\details{
This is the single-cell loader of AbID-data. It will do normalization on control and
mappability and generate multiple handy tracks and metadata.
}
