% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{mabid_correlate}
\alias{mabid_correlate}
\title{General correlation function}
\usage{
mabid_correlate(
  sce,
  blacklist,
  ROI = NULL,
  GRlist_other = NULL,
  assay_to_use = "logcounts",
  bin_size = NULL,
  kernel_size = 4,
  winsorize = 0.9995,
  npc = 7,
  ignore_negatives = F
)
}
\arguments{
\item{sce}{A SCE object}

\item{blacklist}{Ignore these regions}

\item{ROI}{Region of interest to correlate to}

\item{GRlist_other}{add a list of bigwig-GRs to correlate to}

\item{assay_to_use}{assay in sce to use}

\item{bin_size}{bin-size to use}

\item{kernel_size}{Smoothing window}

\item{winsorize}{Winsorize (Replace Extreme Values by Less Extreme Ones)}

\item{npc}{threads to use}

\item{ignore_negatives}{set negative values to zero}
}
\description{
`mabid_correlate()`
}
\details{
Get different correlation-metrics (region and chromosome-wide)
}
