% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_get_track}
\alias{mabid_get_track}
\title{Pull values of a ROI from the SCE.}
\usage{
mabid_get_track(
  sce,
  chrom = "chr1",
  start = 0,
  end = 5e+06,
  assay_to_use = "logcounts",
  metadata = T
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{chrom}{Chromosome of interest}

\item{start}{Start of region}

\item{end}{End of region}

\item{assay_to_use}{Use data from one of the assays the sce. Look in `names(assays(sce))`}

\item{metadata}{Will add all colData. Can slow things down!!!}
}
\value{
A data.table-object
}
\description{
`mabid_get_track()` returns a datatable of the values of a regions of interest
}
\details{
This enables quick querying and subsetting of an SCE
}
\examples{
\dontrun{
data_roi <- m_get_track(sce_SBC001,
                           chrom = 'chr1', start = 0, end = 100e6,
                           assay_to_use = 'OE_control')
}
}
