% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabid_tools.R
\name{mabid_IC}
\alias{mabid_IC}
\title{mabid_IC}
\usage{
mabid_IC(sce, counts_mappability, p = 7)
}
\arguments{
\item{sce}{A SCE object}

\item{counts_mappability}{A named vector of mappability-counts. Names should be in 'chrX:0-1' format}

\item{p}{Number of CPU's to use}
}
\value{
An SCE-object of the RangedSummarizedExperiment-subclass, with an added "IC" colData-column
}
\description{
`mabid_IC()` calculates and adds information content
}
