% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mabid_load_sample_sheet}
\alias{mabid_load_sample_sheet}
\title{mabid_load_sample_sheet}
\usage{
mabid_load_sample_sheet(path)
}
\arguments{
\item{path}{The location of the sample-sheet}
}
\value{
a data.table
}
\description{
Load and parse a sample-sheet
}
