% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_mirrorplot}
\alias{mabid_mirrorplot}
\title{Plot values of a ROI from a SCE as mirrors}
\usage{
mabid_mirrorplot(
  sce,
  chrom = "chr1",
  start = 0,
  end = 5e+07,
  mirror_list,
  tip = T,
  panel_width = 1,
  smooth_kernel = 1,
  assay_to_use = "logcounts",
  GR = NULL,
  ylim = NULL,
  plot_ratio = 1/1.618
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{chrom}{Chromosome of interest}

\item{start}{Start of region}

\item{end}{End of region}

\item{mirror_list}{A named list of named two-length vectors of sample names.}

\item{tip}{Show a tip on how to save the ggplot.}

\item{panel_width}{How wide should the}

\item{smooth_kernel}{Average over N bins}

\item{assay_to_use}{Use data from one of the assays the sce. Look in `names(assays(sce))`}

\item{GR}{add bigwig-like tracks.}

\item{ylim}{Set range of y-axis}

\item{plot_ratio}{Set aspect ratio (default: 1/1.618)}
}
\value{
A Ggplot2-object
}
\description{
`mabid_mirrorplot()` returns generates mirrorplots.
}
\details{
Pretty pretty!
}
\examples{
\dontrun{

mirror_list <- list("H3K36me3" = c('single' = "H3K36me3: singlet",'combined' = "H3K36me3: combined"),
'H3K27me3' = c('single' = "H3K27me3: singlet", 'combined' = "H3K27me3: combined"),
'PolII S5P' = c('single' = 'RNA PolII CTD S5P: singlet','combined' = 'RNA PolII CTD S5P: combined'),
'H3' = c('single' = "H3: singlet", 'combined' = "H3: combined"))
mabid_mirrorplot(sce_SBC001,mirror_list = mirror_list,
                           chrom = 'chr1', start = 0, end = 100e6,
                           assay_to_use = 'logcounts')
}
}
