% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_populationplot}
\alias{mabid_populationplot}
\title{Plot values of a ROI from a SCE as raster, with a population-average on top}
\usage{
mabid_populationplot(
  sce,
  chrom = "chr1",
  start = 2e+07,
  end = 3e+07,
  group_by = NULL,
  assay_to_use = "logcounts",
  ylim = NULL,
  zlim = NULL,
  hline = 0,
  vline = NULL,
  line_colour = "black",
  line_type = 1,
  plot_ratio = 1/10,
  FUN = mean,
  title = NULL,
  subtitle = NULL,
  divpal = F
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{chrom}{Chromosome of interest}

\item{start}{Start of region}

\item{end}{End of region}

\item{group_by}{Select one (!!!) colData-column to split on}

\item{assay_to_use}{Use data from one of the assays the sce. Look in `names(assays(sce))`}

\item{ylim}{Set range of y-axis}

\item{zlim}{Set range of z-axis (= fill intensity)}

\item{hline}{Add horizontal line to upper plot}

\item{vline}{Add vertical line to both plots}

\item{line_colour}{Set line colour}

\item{line_type}{Set line type}

\item{plot_ratio}{Set height-ratios of upper versus lower plot (default is 1/10)}

\item{FUN}{Function to aggregate over. Set to mean}

\item{title}{Title}

\item{subtitle}{Subtitle}

\item{divpal}{Use a divergent colourmap?}
}
\value{
A Ggplot2-object
}
\description{
`mabid_populationplot()` generates a heatmap
}
\details{
Pretty pretty!
}
\examples{
\dontrun{
mabid_populationplot(sce, end = 249250621, assay_to_use = 'counts')
}
}
