% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_rasterplot}
\alias{mabid_rasterplot}
\title{Plot values of a ROI from a SCE as rasterplot}
\usage{
mabid_rasterplot(
  sce,
  chrom = "chr1",
  start = 0,
  end = 5e+06,
  assay_to_use = "logcounts",
  clustering_method = "ward.D2",
  clustering_distance = "canberra",
  divpal = F,
  zlim = NULL,
  asp = 1.618034,
  size = NULL,
  font_size = 10,
  rownames = F,
  annotation_columns = NULL,
  annotation_col = NULL
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{chrom}{Chromosome of interest}

\item{start}{Start of region}

\item{end}{End of region}

\item{assay_to_use}{Use data from one of the assays the sce. Look in `names(assays(sce))`}

\item{clustering_method}{The agglomeration method to be used. This should be
one of "ward.D", "ward.D2", "single",
"complete", "average", "mcquitty", "median" or "centroid".}

\item{clustering_distance}{the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{divpal}{Use a divergent colourmap?}

\item{zlim}{Set range of z-axis (= fill intensity)}

\item{asp}{The aspect-ratio of the plot}

\item{size}{Set size of the heatmap}

\item{font_size}{Set overall font size}

\item{rownames}{Display sample-names?}

\item{annotation_columns}{a vector of column-names of colData(sce)}

\item{annotation_col}{a vector of hex-colours to use}
}
\value{
A Ggplot2-object
}
\description{
`mabid_rasterplot()` generates a heatmap
}
\details{
Pretty pretty!
}
\examples{
\dontrun{
mabid_rasterplot(sce, end = 249250621, assay_to_use = 'counts')
}
}
