% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabid_tools.R
\name{mabid_rebin}
\alias{mabid_rebin}
\title{mabid_rebin}
\usage{
mabid_rebin(sce, bin_size = 20000, threads = 7)
}
\arguments{
\item{sce}{Object: SCE}

\item{bin_size}{Integer: size of a bin in bp}

\item{threads}{Integer: number of CPU's}
}
\value{
An SCE-object of the RangedSummarizedExperiment-subclass (logcounts = TMM-normalised lfc)
}
\description{
`mabid_rebin()` re-bins a SCE
}
\details{
Change the bin-size a sce-object containing MAbID-datasets.
}
