% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_SVD}
\alias{mabid_SVD}
\title{Performd SVD and plot results}
\usage{
mabid_SVD(
  sce,
  svx = "SV_1",
  svy = "SV_2",
  colour_by = NULL,
  shape_by = NULL,
  assay_to_use = "logcounts"
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{svx}{Which SV on the x-axis?}

\item{svy}{Which SV on the y-axis?}

\item{colour_by}{Use a column of the output of `mabid_get_track(sce)` (default: 'sample')}

\item{shape_by}{Use a column of the output of `mabid_get_track(sce)` (default: 'sample')}

\item{assay_to_use}{Use data from one of the assays the sce (default: `logcounts`). Look in `names(assays(sce))`}
}
\value{
A Ggplot2-object
}
\description{
`mabid_SVD()` returns a ggplot2-object
}
\details{
Pretty pretty!
}
\examples{
\dontrun{
mabid_SVD(sce_SBC001)
}
}
