% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mabid.R
\name{mabid_trackplot}
\alias{mabid_trackplot}
\title{Plot values of a ROI from a SCE as barplots.}
\usage{
mabid_trackplot(
  sce,
  chrom = "chr1",
  start = 0,
  end = 5e+06,
  colour_on = "sample",
  kernelLen = 3,
  annot_GRlist = NULL,
  annot_col = NULL,
  assay_to_use = "logcounts",
  GR = NULL,
  ylim = NULL,
  ylim_per_sample = T,
  hline = NULL,
  plot_ratio = 1/10
)
}
\arguments{
\item{sce}{An MAb-ID SCE object}

\item{chrom}{Chromosome of interest}

\item{start}{Start of region}

\item{end}{End of region}

\item{colour_on}{Use a column of the output of `mabid_get_track(sce)` (default: 'sample')}

\item{kernelLen}{(integer) Number of data-points to consider in the kernel.}

\item{annot_GRlist}{a named list of GR's}

\item{annot_col}{Set a standard colour for de annot_GRlist}

\item{assay_to_use}{Use data from one of the assays the sce. Look in `names(assays(sce))`}

\item{GR}{add bigwig-like tracks.}

\item{ylim}{Set range of y-axis}

\item{ylim_per_sample}{Automagically get nice ylims}

\item{hline}{Add a horizontal line at a y-axis value}

\item{plot_ratio}{Set aspect ratio (default: 1/10)}
}
\value{
A Ggplot2-object
}
\description{
`mabid_trackplot()` returns generates barplots, coloured on sample.
}
\details{
Pretty pretty!
}
\examples{
\dontrun{
data_roi <- mabid_get_track(sce_SBC001,
                           chrom = 'chr1', start = 0, end = 100e6,
                           assay_to_use = 'OE_control')
}
}
